+++ src/components/content/notification-handler/notification-handler.cy.js
@@ -1,0 +1,12 @@
+import NotificationHandler from "./notification-handler.vue";
+import { createMount } from "@cypress/support/mount";
+
+const mount = createMount(NotificationHandler);
+
+describe("notification-handler", () => {
+	it("A component is rendered", () => {
+		mount();
+
+		cy.getByData("notification-handler").shouldBeVisible();
+	});
+});