+++ src/components/icon/icon-bell/icon-bell.vue
@@ -1,0 +1,8 @@
+<template>
+	<base-icon>
+		<g fill="none">
+			<path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5" d="M15.75 13.25a2 2 0 0 1-2-2V6.5A4.75 4.75 0 0 0 9 1.75h0A4.75 4.75 0 0 0 4.25 6.5v4.75a2 2 0 0 1-2 2z" />
+			<path fill="currentColor" d="M10.588 15.185A.5.5 0 0 0 10.2 15H7.801a.5.5 0 0 0-.489.603C7.485 16.425 8.18 17 9.001 17s1.516-.575 1.689-1.397a.5.5 0 0 0-.101-.418z" />
+		</g>
+	</base-icon>
+</template>