+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [`form-checkbox-group`](/src/components/form/form-checkbox-group/form-checkbox-group.md)
 - [`star-rating`](/src/components/interaction/star-rating/star-rating.md)
+- [`notification-handler`](/src/components/content/notification-handler/notification-handler.md)
 
 ### Updates